% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs.R
\name{usgs_significant}
\alias{usgs_significant}
\title{Get significant earthquakes in the last 30 days}
\usage{
usgs_significant()
}
\value{
tibble: same columns as usgs_quakes
}
\description{
Get significant earthquakes in the last 30 days
}
