% Generated by build script
\name{water_realtime}
\alias{water_realtime}
\title{Fetch real-time (instantaneous) water data from USGS}
\usage{
water_realtime()
}
\arguments{
\item{sites}{Character vector of USGS site numbers (e.g., "01646500")}
\item{params}{Parameter codes: "00060" = discharge, "00065" = gage height,}
\item{period}{ISO 8601 duration (e.g., "P1D" = past 1 day, "P7D" = past week).}
}
\value{tibble: site, parameter, datetime, value, qualifier, site_name}
\description{Fetch real-time (instantaneous) water data from USGS}
