% Generated by build script
\name{water_site_info}
\alias{water_site_info}
\title{Get site info for a specific USGS monitoring site}
\usage{
water_site_info()
}
\arguments{
\item{site}{USGS site number (e.g., "01646500")}
}
\value{tibble: site_no, site_name, site_type, lat, lon, state_cd,}
\description{Get site info for a specific USGS monitoring site}
