% Generated by build script
\name{water_sites}
\alias{water_sites}
\title{Search for USGS water monitoring sites by state}
\usage{
water_sites()
}
\arguments{
\item{state}{Two-letter state code (e.g., "MD", "CA")}
\item{site_type}{Site type code. Default "ST" (stream). Use "GW" for}
\item{param_code}{Optional parameter code to filter sites that measure}
}
\value{tibble: site_no, site_name, site_type, lat, lon, state_cd,}
\description{Search for USGS water monitoring sites by state}
