% Generated by build script
\name{water_stats}
\alias{water_stats}
\title{Get statistical summaries for a USGS site}
\usage{
water_stats()
}
\arguments{
\item{site}{USGS site number (e.g., "01646500")}
\item{params}{Parameter codes. Default "00060" (discharge).}
\item{stat_type}{Statistic type: "daily" (default), "monthly", "annual"}
}
\value{tibble: site, parameter, month_day, mean, median, count_years}
\description{Get statistical summaries for a USGS site}
